Added missing enumerants, determine enumerations underlying type.

+++ OpenGL.Net/Cl.VERSION_2_0.cs
@@ -215,7 +215,7 @@ namespace OpenCL
 				fixed (IntPtr* p_event = @event)
 				{
 					Debug.Assert(Delegates.pclEnqueueSVMMap != null, "pclEnqueueSVMMap not implemented");
-					retValue = Delegates.pclEnqueueSVMMap(command_queue, blocking_map, flags, svm_ptr, size, num_events_in_wait_list, p_event_wait_list, p_event);
+					retValue = Delegates.pclEnqueueSVMMap(command_queue, blocking_map, (ulong)flags, svm_ptr, size, num_events_in_wait_list, p_event_wait_list, p_event);
 					LogFunction("clEnqueueSVMMap(0x{0}, {1}, {2}, 0x{3}, {4}, {5}, {6}, {7}) = {8}", command_queue.ToString("X8"), blocking_map, flags, svm_ptr.ToString("X8"), size, num_events_in_wait_list, LogValue(event_wait_list), LogValue(@event), retValue);
 				}
 			}
